Remove drag disabiling on mousemove
When using drag delay, the event is cancelled on a series of events, including `mousemove` which seems to be always fired on Chrome Android, even in cases when `touchmove` doesn't fire. This cause the drag to never commence when using the delay property. See https://github.com/RubaXa/Sortable/issues/981